% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_heatmap_dynamic.R
\name{clustering_heatmap_dynamic}
\alias{clustering_heatmap_dynamic}
\title{Plot heat map of clustering of expression value}
\usage{
clustering_heatmap_dynamic(input_matrix, clusters)
}
\arguments{
\item{input_matrix}{Matrix. Expression values.}

\item{clusters}{Named vector of cluster assignment}
}
\value{
ComplexHeatmap
}
\description{
Plot heat map of clustering of expression value
}
