% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_heatmap_hybrid.R
\name{clustering_heatmap_hybrid}
\alias{clustering_heatmap_hybrid}
\title{Plot heat map of clustering of expression value}
\usage{
clustering_heatmap_hybrid(input_matrix, dendrogram, cluster_no, dynamic_labels)
}
\arguments{
\item{input_matrix}{Matrix. Expression values.}

\item{dendrogram}{Dendrogram object.}

\item{cluster_no}{Numeric. Number of clusters.}

\item{dynamic_labels}{Dynamic clustering labels}
}
\value{
ComplexHeatmap
}
\description{
Plot heat map of clustering of expression value
}
