% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_heatmap_static.R
\name{clustering_heatmap_static}
\alias{clustering_heatmap_static}
\title{Plot heat map of clustering of expression value}
\usage{
clustering_heatmap_static(input_matrix, dendrogram, cluster_no)
}
\arguments{
\item{input_matrix}{Matrix. Expression values.}

\item{dendrogram}{Dendrogram object.}

\item{cluster_no}{Numeric. Number of clusters.}
}
\value{
ComplexHeatmap
}
\description{
Plot heat map of clustering of expression value
}
