% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dend_get_merge_height.R
\name{dend_get_merge_height}
\alias{dend_get_merge_height}
\title{Function to get the merge height of a vector of labels in a dendrogram. These are helper functions for dynamic tree cut}
\usage{
dend_get_merge_height(dendrogram, dend_labels)
}
\arguments{
\item{dendrogram}{Dendrogram object}

\item{dend_labels}{Labels of the sub-dendrogram}
}
\value{
Numeric.
}
\description{
Function to get the merge height of a vector of labels in a dendrogram. These are helper functions for dynamic tree cut
}
