% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_data.R
\name{expression_data}
\alias{expression_data}
\title{Retrieve TMM expression data of genes}
\usage{
expression_data(threshold = 1, photoperiodic_only = FALSE, condition_count = 3)
}
\arguments{
\item{threshold}{Filter by expression level. (Sum of TMM reads in each of the three photoperiods)}

\item{photoperiodic_only}{Whether to include only genes that show differential expression between photoperiods}

\item{condition_count}{1, 2 or 3, the number of photoperiods that needs to be detected above threshold}
}
\value{
List of matrices
}
\description{
Retrieve TMM expression data of genes
}
