% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{expression_matrix_photoperiodic}
\alias{expression_matrix_photoperiodic}
\title{Expression of photoperiodic genes in mean TMM calculated by edgeR}
\format{
A data frame with 8293 rows and 18 variables:
}
\usage{
expression_matrix_photoperiodic
}
\description{
Expression of photoperiodic genes in mean TMM calculated by edgeR
}
\keyword{datasets}
