% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gene_identifiers}
\alias{gene_identifiers}
\title{Gene names for the matrix expression_matrix_photoperiodic}
\format{
Character vector
}
\usage{
gene_identifiers
}
\description{
Gene names for the matrix expression_matrix_photoperiodic
}
\keyword{datasets}
