% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_TAIR.R
\name{get_TAIR}
\alias{get_TAIR}
\title{Function to generate a TxDb object from the GenomicFeatures object for getting gene promoters.}
\usage{
get_TAIR(annotation = "description")
}
\arguments{
\item{annotation}{Either "description", "pubmed", "symbols}
}
\value{
Returns a TxDb object (from the GenomicFeatures package)
}
\description{
Function to generate a TxDb object from the GenomicFeatures object for getting gene promoters.
}
