% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gene.R
\name{plot_gene}
\alias{plot_gene}
\title{Extract significantly up- and down-reguluted genes from edgeR results}
\usage{
plot_gene(
  gene,
  tag = "",
  ymax = 4,
  ymin = -4,
  transparency = 1,
  font_size = 7,
  label_size = 3.5,
  label_x = 4,
  label_y = -3
)
}
\arguments{
\item{gene}{Character}

\item{tag}{Character}

\item{ymax}{Numeric}

\item{ymin}{Numeric}

\item{transparency}{Numeric}

\item{font_size}{font size for axis labels}

\item{label_size}{size of gene name annotation}

\item{label_x}{label x position}

\item{label_y}{label y position}
}
\value{
None
}
\description{
Extract significantly up- and down-reguluted genes from edgeR results
}
