% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_aggcluster_dynamic.R
\name{process_aggcluster_dynamic}
\alias{process_aggcluster_dynamic}
\title{Function to}
\usage{
process_aggcluster_dynamic(
  aggcluster_out,
  original_matrix,
  identifiers,
  deepsplit_degree = 1
)
}
\arguments{
\item{aggcluster_out}{output object from AggExCluster}

\item{original_matrix}{matrix}

\item{identifiers}{identifiers}

\item{deepsplit_degree}{1, 2, 3 or 4}
}
\value{
list
}
\description{
Function to
}
