% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_aggcluster_static.R
\name{process_aggcluster_static}
\alias{process_aggcluster_static}
\title{Function to}
\usage{
process_aggcluster_static(aggcluster_out, identifiers, k)
}
\arguments{
\item{aggcluster_out}{output object from AggExCluster}

\item{identifiers}{Character. Name of genes}

\item{k}{number of clusters}
}
\value{
list
}
\description{
Function to
}
