% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rDEI_summary}
\alias{rDEI_summary}
\title{Information on daily expression integral.}
\format{
A data frame with 53940 rows and 10 variables:
\describe{
  \item{geneID}{}
  \item{EQ_DEI}{daily expression integral at equinox photoperiod}
  \item{LD_DEI}{daily expression integral at long day photoperiod}
  \item{SD_DEI}{daily expression integral at short day photoperiod}
  \item{log2_rDEI_LDEQ}{}
  \item{log2_rDEI_SDEQ}{}
  \item{log2_rDEI_SDLD}{}
}
}
\usage{
rDEI_summary
}
\description{
Information on daily expression integral.
}
\keyword{datasets}
