% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gene.R
\name{plot_gene}
\alias{plot_gene}
\title{Extract significantly up- and down-reguluted genes from edgeR results}
\usage{
plot_gene(gene, tag = "", ymax = 4, ymin = -2, transparency = 1)
}
\arguments{
\item{gene}{Character}

\item{tag}{Character}

\item{ymax}{Numeric}

\item{ymin}{Numeric}

\item{transparency}{Numeric}
}
\value{
None
}
\description{
Extract significantly up- and down-reguluted genes from edgeR results
}
